float

+++ cercle.py
@@ -2,6 +2,6 @@
 
 radi = input("Radi de la circumferència? ")
 PI = 3.14
-perimetre = 2*PI*int(radi)
+perimetre = 2*PI*float(radi)
 print("El perímetre de la cirfumferència de radi {} és {}".format(radi,perimetre))
